Necessary classes for crypto trading app
In particular, the CSVReader class contains a self-written csv parser to automatically read to memory data stored in a csv file. 
Another key element are the OrderBookEntry types that identify OrderTypes, specifically buy and sell orders. 
Also included is a matching engine that matches trades input by the user against limit orders that exist in the orderbook.